src/requirements-conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
ale-py=0.7.5=pypi_0
anyio=3.5.0=pypi_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310he2412df_2
astor=0.8.1=pypi_0
asttokens=2.0.5=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
autorom=0.4.2=pypi_0
autorom-accept-rom-license=0.4.2=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
blas=1.0=mkl
bleach=5.0.0=pyhd8ed1ab_0
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=pypi_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=py310hcbf9ad4_0
charset-normalizer=2.0.12=pypi_0
click=8.1.2=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cycler=0.11.0=pypi_0
cython=0.29.28=pypi_0
debugpy=1.6.0=py310h8a704f9_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.4=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
fasteners=0.17.3=pypi_0
flatbuffers=2.0=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.32.0=pypi_0
future=0.18.2=py310h5588dad_5
gast=0.2.2=pypi_0
glfw=2.5.3=pypi_0
google-auth=2.6.5=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
gym=0.23.1=pypi_0
gym-notices=0.0.6=pypi_0
gym-super-mario-bros=7.3.2=pypi_0
h5py=3.6.0=pypi_0
idna=3.3=pypi_0
imageio=2.17.0=pypi_0
importlib-metadata=4.11.3=py310h5588dad_1
importlib_resources=5.7.1=pyhd8ed1ab_0
intel-openmp=2022.0.0=h57928b3_3663
ipykernel=6.13.0=py310hbbfc1a7_0
ipython=8.2.0=py310h5588dad_0
ipython-genutils=0.2.0=pypi_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py310h5588dad_1
jinja2=3.1.1=pyhd8ed1ab_0
joblib=1.1.0=pypi_0
json5=0.9.6=pypi_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter-server=1.16.0=pypi_0
jupyter_client=7.2.2=pyhd8ed1ab_1
jupyter_core=4.9.2=py310h5588dad_0
jupyterlab=3.3.4=pypi_0
jupyterlab-server=2.12.0=pypi_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
keras=2.8.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=pypi_0
libblas=3.9.0=12_win64_mkl
libcblas=3.9.0=12_win64_mkl
libclang=13.0.0=pypi_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=12_win64_mkl
libsodium=1.0.18=h8d14728_1
libuv=1.43.0=h8ffe710_0
libzlib=1.2.11=h8ffe710_1014
lockfile=0.12.2=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.1.1=py310he2412df_1
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py310he2412df_1005
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py310hcf6e17e_0
mpi4py=3.1.3=pypi_0
mujoco-py=2.1.2.14=pypi_0
nbclassic=0.3.7=pypi_0
nbclient=0.6.0=pyhd8ed1ab_0
nbconvert=6.5.0=pyhd8ed1ab_0
nbconvert-core=6.5.0=pyhd8ed1ab_0
nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
nbformat=5.3.0=pyhd8ed1ab_0
nes-py=8.1.8=pypi_0
nest-asyncio=1.5.5=pyhd8ed1ab_0
ninja=1.10.2=h2d74725_1
notebook=6.4.11=pyha770c72_0
notebook-shim=0.1.0=pypi_0
numpy=1.22.3=py310hcae7c84_2
oauthlib=3.2.0=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=3.0.2=h8ffe710_1
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=pypi_0
pandoc=2.18=h57928b3_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.1.0=pypi_0
pip=22.0.4=pyhd8ed1ab_0
progressbar2=4.0.0=pypi_0
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.29=pyha770c72_0
protobuf=3.20.0=pypi_0
psutil=5.9.0=py310he2412df_1
ptan=0.4=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygame=2.1.0=pypi_0
pyglet=1.5.11=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pyparsing=3.0.8=pyhd8ed1ab_0
pyrsistent=0.18.1=py310he2412df_1
python=3.10.4=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python-utils=3.1.0=pypi_0
python_abi=3.10=2_cp310
pytorch=1.10.2=cpu_py310habb1308_0
pytz=2022.1=pypi_0
pywin32=303=pypi_0
pywinpty=2.0.5=py310h00ffb61_1
pyzmq=22.3.0=py310h73ada01_2
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=62.1.0=py310h5588dad_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.2=h8ffe710_0
stable-baselines=2.10.2=pypi_0
stack_data=0.2.0=pyhd8ed1ab_0
tbb=2021.5.0=h2d74725_1
tensorboard=1.15.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5=pypi_0
tensorflow-estimator=1.15.1=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=py310h5588dad_1
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
tornado=6.1=py310he2412df_3
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.9=pypi_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
websocket-client=1.3.2=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=py310h5588dad_0
winpty=0.4.3=4
wrapt=1.14.0=pypi_0
xz=5.2.5=h62dcd97_1
zeromq=4.3.4=h0e60522_1
zipp=3.8.0=pyhd8ed1ab_0
zmq=0.0.0=pypi_0
